nlp-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
spacy==3.7.2
sentence-transformers==2.3.1
numpy==1.26.3
scikit-learn==1.4.0
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
qdrant-client==1.7.3
pyyaml==6.0.1
structlog==24.1.0
torch==2.1.2
pytest==7.4.4
